Render component in chatbox
