+++ packages/frontend-core/src/components/Chatbox/Component/Component.svelte
