+++ packages/frontend-core/src/components/Chatbox/Component/_Form.svelte
