+++ packages/frontend-core/src/components/Chatbox/Component/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Component.svelte"
