--- packages/frontend-core/src/components/agents/index.ts
@@ -1,2 +1,0 @@
-export { default as Component } from "./Component/Component.svelte"
-export { default as AgentChatbox } from "./AgentChatbox.svelte"
